Fix multipart emails handling and payload string decoding
Improved mbox_to_df() to use email.walk() in order to recursively
read all levels of multipart emails and extract the text.

Also, the output used to be just the bytes object converted to str.
Now this is changed to decode it using the latin-1 codex, which is
implicitly used by all messages in the corpus.