Fix regex and docstring of tokenize()
The previous versions were trying to account for the presence of
contractions like "don't", however they are not very important.
Plurals are not going to matter after lemmatization and the usual
forms like "haven't" are bound to be removed in any case since both
words are stopwords.

In addition, the format of words has been restricted a little, so
that it doesn't allow consecutive dashes.