Add word2vec_features() and its helper functions
The function vectorizes the dataset using gensim.Word2Vec. After
creating the vectors for each word in the vocabulary, it removes
the words that do not exist in the vocabulary from the dataset using
filter_vocab_words() and then gets the mean vector for all of the
remaining words using get_mean_vector().