Improve HTML parsing and rerun baseline
Now html_parse() treats inline tags differently in order to avoid
adding extra newlines that would ruin the format and make style
features harder to extract.

Also, the url is extracted from <a> tags in order to achieve better
deduplication of plaintext/html. This is also improved by stripping
all the whitespace before comparing the two halves of the string.